+++ eip_7594/src/lib.rs
@@ -1,10 +1,9 @@
 use std::collections::HashMap;
 
-use anyhow::{anyhow, ensure, Result};
-use c_kzg::{Blob as CKzgBlob, Bytes48, Cell as CKzgCell, KzgProof as CKzgProof, KzgSettings};
+use anyhow::{ensure, Result};
+use c_kzg::{Blob as CKzgBlob, Bytes48, Cell as CKzgCell, KzgProof as CKzgProof};
 use hashing::ZERO_HASHES;
 use helper_functions::predicates::is_valid_merkle_branch;
-use kzg::eip_4844::{load_trusted_setup_string, BYTES_PER_G1, BYTES_PER_G2};
 use num_traits::One as _;
 use sha2::{Digest as _, Sha256};
 use ssz::{ByteVector, ContiguousList, ContiguousVector, SszHash, Uint256};
@@ -22,6 +21,10 @@ use types::{
     preset::Preset,
     traits::{BeaconBlock as _, PostDenebBeaconBlockBody},
 };
+
+use crate::trusted_setup::settings;
+
+mod trusted_setup;
 
 const MAX_BLOBS_PER_BLOCK: u64 = 6;
 
@@ -85,7 +88,7 @@ pub fn verify_kzg_proofs<P: Preset>(data_column_sidecar: &DataColumnSidecar<P>) 
         .map(|(_, row_index)| (row_index, index))
         .unzip();
 
-    let kzg_settings = load_kzg_settings()?;
+    let kzg_settings = settings();
 
     let column = column
         .clone()
@@ -188,7 +191,7 @@ pub fn get_custody_columns(node_id: NodeId, custody_subnet_count: u64) -> Vec<Co
 pub fn compute_extended_matrix(blobs: Vec<CKzgBlob>) -> Result<ExtendedMatrix> {
     let mut extended_matrix: Vec<CKzgCell> = Vec::new();
 
-    let kzg_settings = load_kzg_settings()?;
+    let kzg_settings = settings();
 
     for blob in blobs {
         let cells = *CKzgCell::compute_cells(&blob, &kzg_settings)?;
@@ -205,7 +208,7 @@ fn recover_matrix(
     cells_dict: &HashMap<(BlobIndex, CellID), CKzgCell>,
     blob_count: usize,
 ) -> Result<ExtendedMatrix> {
-    let kzg_settings = load_kzg_settings()?;
+    let kzg_settings = settings();
 
     let mut extended_matrix = Vec::new();
     for blob_index in 0..blob_count {
@@ -239,7 +242,7 @@ pub fn get_data_column_sidecars<P: Preset>(
         let kzg_commitments_inclusion_proof =
             kzg_commitment_inclusion_proof(post_deneb_beacon_block_body);
 
-        let kzg_settings = load_kzg_settings()?;
+        let kzg_settings = settings();
 
         let signed_block_header = SignedBeaconBlockHeader {
             message: beacon_block.to_header(),
@@ -334,24 +337,6 @@ fn kzg_commitment_inclusion_proof<P: Preset>(
     );
 
     proof
-}
-
-fn load_kzg_settings() -> Result<KzgSettings> {
-    let contents = include_str!("../../kzg_utils/src/trusted_setup.txt");
-    let (g1_bytes, g2_bytes) =
-        load_trusted_setup_string(contents).map_err(|error| anyhow!(error))?;
-
-    KzgSettings::load_trusted_setup(
-        &g1_bytes
-            .chunks_exact(BYTES_PER_G1)
-            .map(|chunk| TryInto::<[u8; BYTES_PER_G1]>::try_into(chunk).map_err(Into::into))
-            .collect::<Result<Vec<_>>>()?,
-        &g2_bytes
-            .chunks_exact(BYTES_PER_G2)
-            .map(|chunk| TryInto::<[u8; BYTES_PER_G2]>::try_into(chunk).map_err(Into::into))
-            .collect::<Result<Vec<_>>>()?,
-    )
-    .map_err(|error| anyhow!(error))
 }
 
 #[cfg(test)]
